Made sure init container scripts exec their process.
When stopping containers, Docker sends a SIGTERM to PID 1. However, previously PID 1 was a bash process, which didn't forward the SIGTERM to child processes. Then prevented the containers from stopping gracefully, and where instead killed after 10 seconds. This change makes sure PID 1 is not the bash process, but the actual container process.