Debug slope correction
There was an error in the calculation of the aspect and heading angle, as the original implementation expects angles from -180 to 180, while we were using 0 to 360 
In addition, the aspect angle on EE seems downhill directed, while it should be uphill. 
All before calculating the radians. 

As a result, also the correction factor needs to be multiplied by gamma0 instead of being divided.