add pre-commit to requirements
This lib is used when a new module is created and/or when someone is working with a module created after sepal_ui v_2.3.0. it's providing handy pre-commit check to avoid issues when the repositories are pushed to GitHub (cleaning notebook output, ensuring linting etc...)

It's automatically installed with the [dev] version of sepal_ui but I think it's better to install it manually for regular users.