allow the "show hidden" option in jupyterlab
I think that now I understand how Jupyter server is started. As the server stays alive as long as the instance is running (without option for closing or restarting it) the user is unable to apply custom configuration options (as they require a restart of the server). 

I simply added here the possibility to show hidden file. when dealing with GitHub repositories that's easier if you see .giconfig, .gitignore etc....