force install the custom version of ipyleaflet
As mentioned in https://github.com/12rambau/sepal_ui/pull/902, we cannot change the version of ipyleaflet used in sepal_ui to a branch. on the other hand it's perfectly possible to force intall it from here as we do for earthengine. 

As long tas the applications are still using a loose pin on ipylaflet (which shoould be the case) it will work.

Mentioning @dfguerrerom for the follow-up